moved waveform creation to roles, dependency injection via components
