+++ src/sdc11073/mdib/descriptorcontainers.py
@@ -509,6 +509,10 @@ class DistributionSampleArrayMetricDescriptorContainer(AbstractMetricDescriptorC
                              pm_qnames.DomainUnit,
                              pm_qnames.DistributionRange)
 
+class AbstractOperationDescriptorProtocol(AbstractDescriptorProtocol):
+    OperationTarget: str
+    MaxTimeToFinish: DurationType | None
+
 
 class AbstractOperationDescriptorContainer(AbstractDescriptorContainer):
     """Represents AbstractOperationDescriptor in BICEPS."""
