+++ src/sdc11073/provider/periodicreports.py
@@ -3,7 +3,7 @@ import time
 from collections import namedtuple
 from functools import reduce
 
-from . import intervaltimer
+from sdc11073 import intervaltimer
 from ..loghelper import get_logger_adapter
 
 PeriodicStates = namedtuple('PeriodicStates', 'mdib_version states')
