+++ tests/test_device_waveform.py
@@ -1,13 +1,13 @@
 import time
 import unittest
-
+from decimal import Decimal
 import sdc11073
 from sdc11073.xml_types import pm_types
 from sdc11073.xml_types.dpws_types import ThisModelType, ThisDeviceType
 from sdc11073.mdib import descriptorcontainers as dc
-from sdc11073.mdib.providerwaveform import Annotator
-from sdc11073.provider import waveforms
+from sdc11073.roles.waveformprovider import waveforms
 from sdc11073.pysoap.soapclientpool import SoapClientPool
+from sdc11073.roles.waveformprovider.waveformproviderimpl import GenericWaveformProvider
 from tests import mockstuff
 
 # data that is used in report
@@ -25,10 +25,12 @@ class TestDeviceWaveform(unittest.TestCase):
         self.mdib.descriptions.add_object(desc)
         for h in HANDLES:
             desc = dc.RealTimeSampleArrayMetricDescriptorContainer(handle=h, parent_handle='some_new_handle')
-            desc.SamplePeriod = 0.1
-            desc.unit = pm_types.CodedValue('abc')
+            desc.SamplePeriod = 0.01
+            desc.Unit = pm_types.CodedValue('abc')
+            desc.TechnicalRange.append(pm_types.Range(Decimal(0), Decimal(10)))
             desc.MetricAvailability = pm_types.MetricAvailability.CONTINUOUS
             desc.MetricCategory = pm_types.MetricCategory.MEASUREMENT
+            desc.Resolution = Decimal(0.01)
             self.mdib.descriptions.add_object(desc)
         self.mdib.xtra.mk_state_containers_for_all_descriptors()
 
@@ -39,16 +41,8 @@ class TestDeviceWaveform(unittest.TestCase):
             self.sdc_device.stop_all()
 
     def test_waveformGeneratorHandling(self):
-        waveform_provider = self.mdib.xtra.waveform_provider
-        self.assertIsNotNone(waveform_provider)
-
-        tr = waveforms.TriangleGenerator(min_value=0, max_value=10, waveformperiod=2.0, sampleperiod=0.005)
-        st = waveforms.SawtoothGenerator(min_value=0, max_value=10, waveformperiod=2.0, sampleperiod=0.01)
-        si = waveforms.SinusGenerator(min_value=-8.0, max_value=10.0, waveformperiod=5.0, sampleperiod=0.05)
-
-        waveform_provider.register_waveform_generator(HANDLES[0], tr)
-        waveform_provider.register_waveform_generator(HANDLES[1], st)
-        waveform_provider.register_waveform_generator(HANDLES[2], si)
+        waveform_provider = GenericWaveformProvider(self.mdib, '')
+        waveform_provider.provide_waveforms()
 
         waveform_generators = waveform_provider._waveform_generators
         # first read shall always be empty
@@ -84,6 +78,18 @@ class TestDeviceWaveform(unittest.TestCase):
         self.assertTrue(abs(now - rt_sample_array.determination_time) <= 0.1)
 
     def test_waveformSubscription(self):
+        self._mk_device()
+
+        test_subscription = mockstuff.TestDevSubscription([self.sdc_device.mdib.sdc_definitions.Actions.Waveform],
+                                                          self._soap_client_pool,
+                                                          self.sdc_device.msg_factory)
+        mgr = self.sdc_device.hosted_services.state_event_service.hosting_service.subscriptions_manager
+        mgr._subscriptions.add_object(test_subscription)
+
+        time.sleep(3)
+        self.assertGreater(len(test_subscription.reports), 10)
+
+    def _mk_device(self):
         this_model = ThisModelType(manufacturer='ABCDEFG GmbH',
                                    manufacturer_url='www.abcdefg.com',
                                    model_name='Foobar',
@@ -94,8 +100,11 @@ class TestDeviceWaveform(unittest.TestCase):
                                      firmware_version='0.99',
                                      serial_number='123serial')
 
-        waveform_provider = self.mdib.xtra.waveform_provider
-        self.assertIsNotNone(waveform_provider)
+        wsd = mockstuff.MockWsDiscovery('5.6.7.8')
+        self.sdc_device = sdc11073.provider.SdcProvider(wsd, this_model, this_device, self.mdib)
+        self.sdc_device.start_all()
+
+        waveform_provider = self.sdc_device.waveform_provider
 
         tr = waveforms.TriangleGenerator(min_value=0, max_value=10, waveformperiod=2.0, sampleperiod=0.02)
         st = waveforms.SawtoothGenerator(min_value=0, max_value=10, waveformperiod=2.0, sampleperiod=0.02)
@@ -105,19 +114,8 @@ class TestDeviceWaveform(unittest.TestCase):
         waveform_provider.register_waveform_generator(HANDLES[1], st)
         waveform_provider.register_waveform_generator(HANDLES[2], si)
 
-        annotator = Annotator(annotation=pm_types.Annotation(pm_types.CodedValue('a', 'b')),
-                              trigger_handle=HANDLES[2],
-                              annotated_handles=[HANDLES[0], HANDLES[1], HANDLES[2]])
-        waveform_provider.register_annotation_generator(annotator)
+        waveform_provider.add_annotation_generator(pm_types.CodedValue('a', 'b'),
+                                                   trigger_handle=HANDLES[2],
+                                                   annotated_handles=[HANDLES[0], HANDLES[1], HANDLES[2]]
+                                                   )
 
-        wsd = mockstuff.MockWsDiscovery('5.6.7.8')
-        self.sdc_device = sdc11073.provider.SdcProvider(wsd, this_model, this_device, self.mdib)
-        self.sdc_device.start_all()
-        test_subscription = mockstuff.TestDevSubscription([self.sdc_device.mdib.sdc_definitions.Actions.Waveform],
-                                                          self._soap_client_pool,
-                                                          self.sdc_device.msg_factory)
-        mgr = self.sdc_device.hosted_services.state_event_service.hosting_service.subscriptions_manager
-        mgr._subscriptions.add_object(test_subscription)
-
-        time.sleep(3)
-        self.assertGreater(len(test_subscription.reports), 10)
